requirements.txt
requests==2.27.1
python-dotenv==0.19.2
selenium~=4.1.0
pandas==1.4.1
xlwings~=0.26.3
openpyxl~=3.0.9